refactor(connector): fix
